# Optimize deserialize -- up to 5x faster