--- a/axlearn/common/checkpointer_test.py
+++ b/axlearn/common/checkpointer_test.py
@@ -31,6 +31,7 @@ from jax import numpy as jnp
 from jax.experimental import mesh_utils
 from jax.experimental.array_serialization import serialization as array_serialization
 
+from axlearn.common import array_serialization as axlearn_serialization
 from axlearn.common import file_system as fs
 from axlearn.common import serialization, test_utils, utils
 from axlearn.common.array_serialization import BoundedDataShardedAsyncCheckpointManager
@@ -1159,16 +1160,23 @@ class TensorStoreStateStorageTest(test_utils.TestCase):
         with self.assertRaisesRegex(RuntimeError, "cannot schedule new futures after shutdown"):
             storage._executor.submit(worker)
 
-    @parameterized.parameters(jnp.float32, jnp.bfloat16, jnp.int32, jnp.int16)
-    def test_save_and_restore_from_dir(self, restore_floats_as: jnp.dtype):
+    @parameterized.product(
+        restore_floats_as=[jnp.float32, jnp.bfloat16, jnp.int32, jnp.int16],
+        premaped_buffer_size=[1024 * 1024 * 1024, 1],
+    )
+    def test_save_and_restore_from_dir(
+        self, restore_floats_as: jnp.dtype, premaped_buffer_size: int
+    ):
         mesh_shape = (1, 1)
         if not test_utils.is_supported_mesh_shape(mesh_shape):
             return
 
         def make_state(float_dtype):
-            return dict(x=jnp.zeros([], dtype=jnp.int32), y=jnp.ones([2], dtype=float_dtype))
+            return dict(x=jnp.zeros([1024], dtype=jnp.int32), y=jnp.ones([2], dtype=float_dtype))
 
-        with _mesh(mesh_shape):
+        with _mesh(mesh_shape), mock.patch.object(
+            axlearn_serialization, "_get_premapped_buffer_size", lambda: premaped_buffer_size
+        ):
             state = make_state(float_dtype=jnp.float32)
             storage = TensorStoreStateStorage.default_config().instantiate()
             with tempfile.TemporaryDirectory() as root_dir: