This improved end-to-end loading time from 100s+ to 25s for a single v5p-8 loading a 70B ckpt in bf16.

It also improves the ckpt load time for GCS too, from ~120s to ~75s.

Ablations:

Baseline: time dominated by numpy `zeros` and `astype`
![image](https://github.com/user-attachments/assets/d784b961-0e05-4bcb-a540-604c44cba554)

With optimization 1 and 2 (no np.zeros and astype), and parallelize all map dma buffers (this consumes ~150G additional memory!). Note that this method leads to high variance in load time. Worse case load time could be 150s+, and best cases would be around ~45s
![image](https://github.com/user-attachments/assets/2e3d724e-fcf6-4643-90c5-031ed5dfa6c6)

Sequential linearization with 16GB pre-mapped buffer (**this PR**):
![image](https://github.com/user-attachments/assets/220612cb-6db9-4c9c-a3ac-78a1ab904d51)

Concurrent linearization with 32GB pre-mapped buffer (**this PR**):

By controlling the pre-mapped buffer size, one can improve performance by increasing the parallelization of linearization operations.
![image](https://github.com/user-attachments/assets/e8ae6bcb-7231-421b-bc87-00a1ee4fe4fe)